baackend/requirement.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
SQLAlchemy==2.0.25
PyJWT==2.8.0
Werkzeug==3.0.1
